+++ src/main/java/leets/leenk/domain/feed/application/usecase/FeedUsecase.java
@@ -34,7 +34,7 @@ import leets.leenk.domain.media.application.mapper.MediaMapper;
 import leets.leenk.domain.media.domain.entity.Media;
 import leets.leenk.domain.media.domain.service.MediaGetService;
 import leets.leenk.domain.media.domain.service.MediaSaveService;
-import leets.leenk.domain.notification.application.usecase.NotificationUsecase;
+import leets.leenk.domain.notification.application.usecase.FeedNotificationUsecase;
 import leets.leenk.domain.user.domain.entity.User;
 import leets.leenk.domain.user.domain.entity.UserBlock;
 import leets.leenk.domain.user.domain.service.NotionDatabaseService;
@@ -71,7 +71,7 @@ public class FeedUsecase {
     private final ReactionGetService reactionGetService;
     private final ReactionSaveService reactionSaveService;
 
-    private final NotificationUsecase notificationUsecase;
+    private final FeedNotificationUsecase feedNotificationUsecase;
 
     private final FeedMapper feedMapper;
     private final MediaMapper mediaMapper;
@@ -118,8 +118,8 @@ public class FeedUsecase {
         List<LinkedUser> linkedUsers = getLinkedUsers(author, request.userId(), feed);
         linkedUserSaveService.saveAll(linkedUsers);
 
-        notificationUsecase.saveNewFeedNotification(feed);
-        notificationUsecase.saveTagNotification(feed, linkedUsers);
+        feedNotificationUsecase.saveNewFeedNotification(feed);
+        feedNotificationUsecase.saveTagNotification(feed, linkedUsers);
     }
 
     private List<LinkedUser> getLinkedUsers(User author, List<Long> userIds, Feed feed) {
@@ -147,7 +147,7 @@ public class FeedUsecase {
         long previousReactionCount = reaction.getReactionCount();
 
         feedUpdateService.updateTotalReaction(feed, reaction, feed.getUser(), request.reactionCount());
-        notificationUsecase.saveFirstReactionNotification(reaction);
+        feedNotificationUsecase.saveFirstReactionNotification(reaction);
 
         long updatedReactionCount = previousReactionCount + request.reactionCount();
         notifyIfReachedReactionMilestone(previousReactionCount, updatedReactionCount, feed);
@@ -252,7 +252,7 @@ public class FeedUsecase {
 
         for (long milestone : milestones) {
             if (previous < milestone && current >= milestone) {
-                notificationUsecase.saveReactionCountNotification(feed, milestone);
+                feedNotificationUsecase.saveReactionCountNotification(feed, milestone);
             }
         }
     }
